Fixed up slick to match WKT and WKB package renames.
