+++ slick/src/main/scala/geotrellis/slick/PostGisSupport.scala
@@ -27,15 +27,17 @@
 
 package geotrellis.slick
 
+import geotrellis.vector._
+import geotrellis.vector.io.wkb._
+import geotrellis.vector.io.wkt._
+
+
 import scala.slick.driver.JdbcDriver
 import scala.slick.lifted.Column
 import scala.reflect.ClassTag
 import scala.slick.ast.{ScalaBaseType}
 import scala.slick.jdbc.{PositionedResult, PositionedParameters}
 import java.sql._
-
-import geotrellis.vector._
-import geotrellis.vector.io._
 
 /** 
  * This class provides column types and extension methods to work with Geometry columns in PostGIS.
