+++ slick/src/test/scala/geotrellis/slick/Data.scala
@@ -20,7 +20,7 @@ package geotrellis.slick
 
 import com.vividsolutions.jts.{geom => jts}
 import geotrellis.vector._
-import geotrellis.vector.io._
+import geotrellis.vector.io.wkt._
 
 object util {
